Fixed Product methods

+++ BestBuyCRUDDatabaseConsole/IProductRepository.cs
@@ -5,7 +5,7 @@ namespace BestBuyCRUDDatabaseConsole
     internal interface IProductRepository
     {
         IEnumerable<Products> GetProducts();
-        void CreateProducts(string createProducts, double productPrice, int productCategory);
+        void CreateProducts(string createProducts, decimal productPrice, int productCategory);
         void UpdateProducts(string updateProducts);
         void DeleteProducts();
